joined convolutionalized models

+++ convolutionalize.py
@@ -1,4 +1,4 @@
-from keras.layers import GlobalAveragePooling2D, Dense, Dropout, Input, Conv2D, AveragePooling2D
+from keras.layers import Conv2D, AveragePooling2D
 from keras.models import Model
 from keras.models import model_from_json
 from PIL import Image
@@ -26,38 +26,47 @@ def prepare_str_file_architecture_syntax(filepath):
     return model_str
 
 
-model = model_from_json(prepare_str_file_architecture_syntax("trained_models/top5_vgg16_acc77_2017-12-24/vgg16_architecture_2017-12-23_22-53-03.json"))
-model.load_weights("trained_models/top5_vgg16_acc77_2017-12-24/vgg16_ft_weights_acc0.78_e15_2017-12-23_22-53-03.hdf5")
-print("IMPORTED MODEL")
-model.summary()
+def load_VGG16(architecture_path, weigths_path, debug=True):
+    model = model_from_json(prepare_str_file_architecture_syntax(architecture_path))
+    model.load_weights(weigths_path)
+    if debug:
+        print("IMPORTED MODEL")
+        model.summary()
 
-p_dim = model.get_layer("global_average_pooling2d_1").input_shape  # None,7,7,512
-out_dim = model.get_layer("output_layer").get_weights()[1].shape[0]  # None,101
-W, b = model.get_layer("output_layer").get_weights()
-print("weights old shape", W.shape, "values", W)
-print("biases old shape", b.shape, "values", b)
+    p_dim = model.get_layer("global_average_pooling2d_1").input_shape
+    out_dim = model.get_layer("output_layer").get_weights()[1].shape[0]
+    W, b = model.get_layer("output_layer").get_weights()
 
-weights_shape = (1, 1, p_dim[3], out_dim)
-print("weights new shape", weights_shape)
+    weights_shape = (1, 1, p_dim[3], out_dim)
 
-new_W = W.reshape(weights_shape)
+    if debug:
+        print("weights old shape", W.shape, "values", W)
+        print("biases old shape", b.shape, "values", b)
+        print("weights new shape", weights_shape)
 
-last_pool_layer = model.get_layer("block5_pool")
-last_pool_layer.outbound_nodes = []
-model.layers.pop()
-model.layers.pop()
+    W = W.reshape(weights_shape)
 
-for i, l in enumerate(model.layers):
-    print(i, ":", l.name)
+    last_layer = model.get_layer("block5_pool")
+    last_layer.outbound_nodes = []
+    model.layers.pop()
+    model.layers.pop()
 
-x = AveragePooling2D(pool_size=(7, 7))(last_pool_layer.output)
+    if debug:
+        for i, l in enumerate(model.layers):
+            print(i, ":", l.name)
 
-x = Conv2D(101, (1, 1), strides=(1, 1), activation='softmax', padding='valid', weights=[new_W, b])(x)
+    return model, last_layer, W, b
 
-model = Model(inputs=model.get_layer("input_1").input, outputs=x)
+baseVGG16_1, last_layer, W, b = load_VGG16("trained_models/top5_vgg16_acc77_2017-12-24/vgg16_architecture_2017-12-23_22-53-03.json", "trained_models/top5_vgg16_acc77_2017-12-24/vgg16_ft_weights_acc0.78_e15_2017-12-23_22-53-03.hdf5")
+x = AveragePooling2D(pool_size=(7, 7), strides=(1, 1))(last_layer.output)
+x = Conv2D(101, (1, 1), strides=(1, 1), activation='softmax', padding='valid', weights=[W, b])(x)
+overlap_fcnVGG16 = Model(inputs=baseVGG16_1.input, outputs=x)
 
-print("CONVOLUTIONALIZATED MODEL")
-model.summary()
+
+baseVGG16_2, last_layer, W, b = load_VGG16("trained_models/top5_vgg16_acc77_2017-12-24/vgg16_architecture_2017-12-23_22-53-03.json", "trained_models/top5_vgg16_acc77_2017-12-24/vgg16_ft_weights_acc0.78_e15_2017-12-23_22-53-03.hdf5")
+x = AveragePooling2D(pool_size=(7, 7))(last_layer.output)
+x = Conv2D(101, (1, 1), strides=(1, 1), activation='softmax', padding='valid', weights=[W, b])(x)
+upsample_fcnVGG16 = Model(inputs=baseVGG16_2.input, outputs=x)
 
 
 def idx_to_class_name(idx):
@@ -65,50 +74,45 @@ def idx_to_class_name(idx):
         class_labels = [line.strip('\n') for line in file.readlines()]
     return class_labels[idx]
 
-def save_map(heatmap, resultfname, is_input_img=False, grid=True):
+def save_map(heatmap, resultfname, tick_interval=None, is_input_img=False):
     if is_input_img:
         pil_input = Image.open(heatmap)
-        pil_input = pil_input.resize(img_size)
+        # pil_input = pil_input.resize(img_size)
         imgarray = np.asarray(pil_input)
     else:
         pixels = 255 * (1.0 - heatmap)
         image = Image.fromarray(pixels.astype(np.uint8), mode='L')
-        image = image.resize(img_size)
+        # image = image.resize(img_size)
         imgarray = np.asarray(image)
 
     fig = plt.figure()
     ax = fig.add_subplot(1, 1, 1)
     # fig.subplots_adjust(left=0, right=1, bottom=0, top=1)
 
-    myInterval = 224.
-    loc = plticker.MultipleLocator(base=myInterval)
-    ax.xaxis.set_major_locator(loc)
-    ax.yaxis.set_major_locator(loc)
-
-    if grid:
-        ax.grid(which='major', axis='both', linestyle='-')
+    if tick_interval:
+        myInterval = tick_interval
+        loc = plticker.MultipleLocator(base=myInterval)
+        ax.xaxis.set_major_locator(loc)
+        ax.yaxis.set_major_locator(loc)
+        ax.grid(which='both', axis='both', linestyle='-')
+        nx = abs(int(float(ax.get_xlim()[1] - ax.get_xlim()[0]) / float(myInterval)))
+        ny = abs(int(float(ax.get_ylim()[1] - ax.get_ylim()[0]) / float(myInterval)))
+        for jj in range(ny):
+            y = myInterval / 2 + jj * myInterval
+            for ii in range(nx):
+                x = myInterval / 2. + float(ii) * myInterval
+                ax.text(x, y, '{:d}'.format((ii + jj * nx) + 1), color='tab:blue', ha='center', va='center')
 
     if is_input_img:
         ax.imshow(imgarray)
     else:
         ax.imshow(imgarray, cmap='Greys_r', interpolation='none')
 
-    nx = abs(int(float(ax.get_xlim()[1] - ax.get_xlim()[0]) / float(myInterval)))
-    ny = abs(int(float(ax.get_ylim()[1] - ax.get_ylim()[0]) / float(myInterval)))
-    for jj in range(ny):
-        y = myInterval / 2 + jj * myInterval
-        for ii in range(nx):
-            x = myInterval / 2. + float(ii) * myInterval
-            ax.text(x, y, '{:d}'.format((ii + jj * nx) + 1), color='tab:blue', ha='center', va='center')
-
     fig.savefig(resultfname)
-    plt.clf()
+    plt.close()
 
-max_upsampling_factor = 6
-min_upsampling_factor = 1
-top_n_show = 3
-threshold_accuracy_stop = 0.80
 
+# dataset-ethz101food/test/apple_pie/101251.jpg
 # "dataset-ethz101food/train/cup_cakes/13821.jpg"
 # "dataset-ethz101food/train/cannoli/1163058.jpg"
 # "dataset-ethz101food/train/apple_pie/68383.jpg"
@@ -122,29 +126,33 @@ threshold_accuracy_stop = 0.80
 #  train/cup_cakes/9256.jpg -> potrebbe dare buoni frutti con upsampling anche a 8
 #  train/cup_cakes/451074.jpg -> molto difficile da trovare ed è lontano ( sopra il bus)
 #  train/cup_cakes/1265596.jpg -> cibo in più posti
-input_class = "cannoli"
-input_instance = "1706697"
-input_set = "test"
+input_set = "train"
+input_class = "cup_cakes"
+input_instance = "46500"
 input_filename = "dataset-ethz101food/" + input_set + "/" + input_class + "/" + input_instance + ".jpg"
 
+threshold_accuracy_stop = 0.80
+max_upsampling_factor = 7
+min_upsampling_factor = 1
+top_n_show = 5
+
+model = overlap_fcnVGG16
+
 if (os.path.exists(input_filename)):
-    for upsampling_factor in range (min_upsampling_factor, max_upsampling_factor + 1):
-        img_size = (224 * upsampling_factor, 224 * upsampling_factor)
-
-        resultdir = os.path.join(os.getcwd(), "results", input_class + "_" + input_instance, "upsampled" + str(upsampling_factor) + "_heatmaps")
-
-        input_image = image.load_img(input_filename, target_size=img_size)
+    if model == overlap_fcnVGG16:
+        input_image = image.load_img(input_filename)
         input_image = image.img_to_array(input_image)
         input_image_expandedim = np.expand_dims(input_image, axis=0)
         input_preprocessed_image = preprocess_input(input_image_expandedim)
+
         preds = model.predict(input_preprocessed_image)
+        print("input img shape (height, width)", input_image.shape, "preds shape", preds.shape)
 
         heatmaps_values = [preds[0, :, :, i] for i in range(101)]
+        max_heatmaps = np.amax(heatmaps_values, axis=(1, 2))
+        top_n_idx = np.argsort(max_heatmaps)[-top_n_show:][::-1]
 
-        max_heatmaps = np.amax(heatmaps_values, axis=(1,2))
-
-        top_n_idx = np.argsort(max_heatmaps)[-3:][::-1]
-
+        resultdir = os.path.join(os.getcwd(), "results", input_class + "_" + input_instance + "_standard-heatmaps")
         if (os.path.isdir(resultdir)):
             print("Deleting older version of the folder " + resultdir)
             shutil.rmtree(resultdir)
@@ -158,11 +166,40 @@ if (os.path.exists(input_filename)):
             save_map(heatmaps_values[idx], resultfname)
             print("heatmap saved at", resultfname)
 
-        if (max_heatmaps[top_n_idx[0]] >= threshold_accuracy_stop):
-            print("Upsampling step " + str(upsampling_factor) + " finished -> accuracy threshold stop detected (accuracy: " + str(max_heatmaps[top_n_idx[0]]) + ")\n")
-            # break
-        else:
-            print("Upsampling step " + str(upsampling_factor) + " finished -> low accuracy, continuing... (accuracy: " + str(max_heatmaps[top_n_idx[0]]) + ")\n")
+    else:
+        for upsampling_factor in range (min_upsampling_factor, max_upsampling_factor):
+            img_size = (255 * upsampling_factor, 255 * upsampling_factor)
+            input_image = image.load_img(input_filename, target_size=img_size)
+            input_image = image.img_to_array(input_image)
+            input_image_expandedim = np.expand_dims(input_image, axis=0)
+            input_preprocessed_image = preprocess_input(input_image_expandedim)
+
+            preds = model.predict(input_preprocessed_image)
+            print("input img shape (height, width)", input_image.shape, "preds shape", preds.shape)
+
+            heatmaps_values = [preds[0, :, :, i] for i in range(101)]
+            max_heatmaps = np.amax(heatmaps_values, axis=(1,2))
+            top_n_idx = np.argsort(max_heatmaps)[-top_n_show:][::-1]
+
+            resultdir = os.path.join(os.getcwd(), "results", input_class + "_" + input_instance, "upsampled" + str(upsampling_factor) + "-heatmaps")
+            if (os.path.isdir(resultdir)):
+                print("Deleting older version of the folder " + resultdir)
+                shutil.rmtree(resultdir)
+            os.makedirs(resultdir)
+
+            save_map(input_filename, os.path.join(resultdir, input_class + "_" + input_instance + ".jpg"), tick_interval=224, is_input_img=True)
+            for i, idx in enumerate(top_n_idx):
+                name_class = idx_to_class_name(idx)
+                print("Top", i, "category is: id", idx, ", name", name_class)
+                resultfname = os.path.join(resultdir, str(i + 1) + "_" + name_class + "_acc" + str(max_heatmaps[idx]) + ".jpg")
+                save_map(heatmaps_values[idx], resultfname, tick_interval=224)
+                print("heatmap saved at", resultfname)
+
+            if (max_heatmaps[top_n_idx[0]] >= threshold_accuracy_stop):
+                print("Upsampling step " + str(upsampling_factor) + " finished -> accuracy threshold stop detected (accuracy: " + str(max_heatmaps[top_n_idx[0]]) + ")\n")
+                break
+            else:
+                print("Upsampling step " + str(upsampling_factor) + " finished -> low accuracy, continuing... (accuracy: " + str(max_heatmaps[top_n_idx[0]]) + ")\n")
 
         # ---------------------------------------------------------------------------
         # wrong way to get the most probable category
@@ -170,5 +207,3 @@ if (os.path.exists(input_filename)):
         # idx_classmax = np.argmax(summed_heatmaps).astype(int)
 else:
     print ("The specified image " + input_filename + " does not exist")
-
-plt.close('all')
